[core] changed dot.NumberOfTicks and dot.TickCount from int to int32 [mage] Living Bomb's dot application can now miss, and its splash damage scaling corrected; multiple applications are no longer supported, they were both bugged and unused [mage] Ignite is now procced by Living Bomb's crit ticks; cleaned up & simplified [druid] Moonfire ticks can no longer miss [druid] Nature's Grace is now additionally procced by Starfall and Moonfire

+++ sim/priest/mind_flay.go
@@ -12,11 +12,11 @@ import (
 // TODO Mind Flay (48156) now "periodically triggers" Mind Flay (58381), probably to allow haste to work.
 //
 //	The first never deals damage, so the latter should probably be used as ActionID here.
-func (priest *Priest) MindFlayActionID(numTicks int) core.ActionID {
-	return core.ActionID{SpellID: 48156, Tag: int32(numTicks)}
+func (priest *Priest) MindFlayActionID(numTicks int32) core.ActionID {
+	return core.ActionID{SpellID: 48156, Tag: numTicks}
 }
 
-func (priest *Priest) newMindFlaySpell(numTicks int) *core.Spell {
+func (priest *Priest) newMindFlaySpell(numTicks int32) *core.Spell {
 	baseCost := priest.BaseMana * 0.09
 
 	channelTime := time.Second * time.Duration(numTicks)
@@ -75,7 +75,7 @@ func (priest *Priest) newMindFlaySpell(numTicks int) *core.Spell {
 	})
 }
 
-func (priest *Priest) newMindFlayDot(numTicks int) *core.Dot {
+func (priest *Priest) newMindFlayDot(numTicks int32) *core.Dot {
 	target := priest.CurrentTarget
 
 	miseryCoeff := 0.257 * (1 + 0.05*float64(priest.Talents.Misery))
@@ -95,7 +95,7 @@ func (priest *Priest) newMindFlayDot(numTicks int) *core.Dot {
 	return core.NewDot(core.Dot{
 		Spell: priest.MindFlay[numTicks],
 		Aura: target.RegisterAura(core.Aura{
-			Label:    "MindFlay-" + strconv.Itoa(numTicks) + "-" + strconv.Itoa(int(priest.Index)),
+			Label:    "MindFlay-" + strconv.Itoa(int(numTicks)) + "-" + strconv.Itoa(int(priest.Index)),
 			ActionID: priest.MindFlayActionID(numTicks),
 		}),
